ajustes generales

+++ src/controllers/Categoria.controller.js
@@ -1,9 +1,22 @@
+import { literal } from "sequelize";
 import { Categoria } from "../models/Categoria.js"
+import { UserAdministrador } from "../models/UserAdministrador.js";
+import { Producto } from "../models/Producto.js";
 
 export const crearCategoria = async (req, res) => {
     try {
         const body = req.body
-        const categoria = await Categoria.create({ idTienda: body.idTienda, nombre: body.nombre, createdBy: 1 });
+        const userAdmin = await UserAdministrador.findOne({
+            attributes: [
+                'id',
+                [literal(`(SELECT tiendas.id FROM tiendas WHERE tiendas.idUserAdministrador = usersAdministradores.id)`), 'idTienda']
+            ],
+            where: {
+                idUser: req.user.id
+            },
+            raw: true
+        });
+        const categoria = await Categoria.create({ idTienda: userAdmin.idTienda, nombre: body.nombre, createdBy: req.user.id });
         return res.status(200).json({ message: 'Categoria guardada con exito', categoria })
     } catch (error) {
         console.log(error)
@@ -14,14 +27,15 @@ export const crearCategoria = async (req, res) => {
 export const actualizarCategoria = async (req, res) => {
     try {
         const body = req.body
+        const params = req.params
         await Categoria.update(
             {
                 nombre: body.nombre,
-                updatedBy: 1
+                updatedBy: req.user.id
             },
             {
                 where: {
-                    id: body.id
+                    id: params.idCategoria,
                 },
             },
         );
@@ -34,10 +48,18 @@ export const actualizarCategoria = async (req, res) => {
 
 export const eliminarCategoria = async (req, res) => {
     try {
-        const query = req.query
+        const params = req.params
+        const categoria = await Producto.count({
+            where: {
+                idCategoria: params.idCategoria,
+            },
+        })
+        if (categoria > 0) {
+            return res.status(400).json({ message: 'No se puede eliminar la categoria porque tiene productos asociados' })
+        }
         await Categoria.destroy({
             where: {
-                id: query.id,
+                id: params.idCategoria,
             },
         });
         return res.status(200).json({ message: 'Categoria eliminada con exito' })
@@ -49,7 +71,18 @@ export const eliminarCategoria = async (req, res) => {
 
 export const obtenerCategoria = async (req, res) => {
     try {
-        const query = req.query
+        const userAdmin = await UserAdministrador.findOne({
+            attributes: [
+                'id',
+                [literal(`(SELECT tiendas.id FROM tiendas WHERE tiendas.idUserAdministrador = usersAdministradores.id)`), 'idTienda']
+            ],
+            where: {
+                idUser: req.user.id
+            },
+            raw: true
+        });
+
+
         const categoria = await Categoria.findAll({
             attributes: [
                 'id',
@@ -58,7 +91,7 @@ export const obtenerCategoria = async (req, res) => {
                 'estado'
             ],
             where: {
-                idTienda: query.idTienda,
+                idTienda: userAdmin.idTienda
             },
         })
         return res.status(200).json({ message: 'Categoria obtenida con exito', data: categoria })
@@ -70,32 +103,32 @@ export const obtenerCategoria = async (req, res) => {
 
 export const cambiarEstadoCategoria = async (req, res) => {
     try {
-        const body = req.body
+        const params = req.params
         const categoria = await Categoria.findOne({
             where: {
-                id: body.id,
+                id: params.idCategoria,
             },
         })
 
         if (categoria.estado == 0) {
             await Categoria.update({
                 estado: 1,
-                updatedBy: 1
+                updatedBy: req.user.idCategoria
             },
                 {
                     where: {
-                        id: body.id
+                        id: params.idCategoria
                     },
                 },
             )
         } else {
             await Categoria.update({
                 estado: 0,
-                updatedBy: 1
+                updatedBy: req.user.idCategoria
             },
                 {
                     where: {
-                        id: body.id
+                        id: params.idCategoria
                     },
                 },
             )
